testPythonPkg reference counts

+++ src/org/ds/python/toxFunctions.groovy
@@ -30,10 +30,11 @@ def getToxTestsParallel(args = [:]){
     def preRunClosure = args['beforeRunning']
     def retries = args.containsKey('retry') ? args.retry : 1
     def dockerRunArgs = args.get('dockerRunArgs', '')
+
     script{
         def envs
         def originalNodeLabel
-        def dockerImageName = "${currentBuild.fullProjectName}:tox".replaceAll("-", "").replaceAll('/', "_").replaceAll(' ', "").toLowerCase()
+        def dockerImageName = "${currentBuild.fullProjectName}_tox_${UUID.randomUUID().toString()}".replaceAll("-", "").replaceAll('/', "_").replaceAll(' ', "").toLowerCase()
         retry(retries){
             node(label){
                 originalNodeLabel = env.NODE_NAME
@@ -42,30 +43,36 @@ def getToxTestsParallel(args = [:]){
                 dockerImage.inside{
                     envs = getToxEnvs()
                 }
+                def untaggingScript = "docker image rm --no-prune ${dockerImage.imageName()}"
                 if(isUnix()){
                     sh(
                         label: "Untagging Docker Image used to run tox",
-                        script: "docker image rm --no-prune ${dockerImage.imageName()}"
+                        script: untaggingScript
                     )
                 } else {
                     bat(
                         label: "Untagging Docker Image used to run tox",
-                        script: "docker image rm --no-prune ${dockerImage.imageName()}"
+                        script: untaggingScript
                     )
                 }
             }
         }
         echo "Found tox environments for ${envs.join(', ')}"
+        def hosts = [:]
         def jobs = envs.collectEntries({ tox_env ->
             def jenkinsStageName = "${envNamePrefix} ${tox_env}"
             [jenkinsStageName,{
                 retry(retries){
                     node(label){
+                        if (!hosts.containsKey(env.NODE_NAME)){
+                            hosts[env.NODE_NAME] = 0
+                        }
                         ws{
                             checkout scm
                             def dockerImageForTesting = docker.build(dockerImageName, "-f ${dockerfile} ${dockerArgs} . ")
                             try{
                                 dockerImageForTesting.inside(dockerRunArgs){
+                                    hosts[env.NODE_NAME] += 1
                                     if(preRunClosure != null){
                                         preRunClosure()
                                     }
@@ -88,32 +95,29 @@ def getToxTestsParallel(args = [:]){
                                     )
                                 }
                             } finally {
-                                if(isUnix()){
-                                    def runningContainers = sh(
-                                                               script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}",
-                                                               returnStdout: true,
-                                                               )
-                                    if (!runningContainers?.trim()) {
-                                        sh(
-                                            label: "Untagging Docker Image used to run tox",
-                                            script: "docker image rm --no-prune ${dockerImageForTesting.imageName()}",
-                                            returnStatus: true
-                                        )
+                                hosts[env.NODE_NAME] -= 1
+                                if (hosts[env.NODE_NAME] == 0){
+                                    if(isUnix()){
+                                        def runningContainers = sh(
+                                                                   script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}",
+                                                                   returnStdout: true,
+                                                                   )
+                                        if (!runningContainers?.trim()) {
+                                            sh(
+                                                label: "Untagging Docker Image used to run tox",
+                                                script: "docker image rm --no-prune ${dockerImageForTesting.imageName()}",
+                                                returnStatus: true
+                                            )
+                                        }
+                                        sh(script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}", returnStatus: true)
+                                    } else {
+                                        def powershellScript ="${env.WORKSPACE_TMP}/cleanupUnusedDockerImage.ps1"
+                                        if (! fileExists(powershellScript) ){
+                                            def scriptContent = libraryResource 'org/ds/cleanupUnusedDockerImage.ps1'
+                                            writeFile file: powershellScript, text: scriptContent
+                                        }
+                                        powershell(script: "${powershellScript} ${dockerImageForTesting.imageName()}", returnStatus: true)
                                     }
-                                    sh(script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}", returnStatus: true)
-                                } else {
-                                    def runningContainers = powershell(
-                                                    returnStdout: true,
-                                                    script: "docker ps --no-trunc --filter \"ancestor=${dockerImageForTesting.imageName()}\" --format \"{{.Names}}\""
-                                                    )
-                                    if (!runningContainers?.trim()) {
-                                        powershell(
-                                            label: "Untagging Docker Image used to run tox",
-                                            script: "docker image rm --no-prune ${dockerImageForTesting.imageName()}",
-                                            returnStatus: true
-                                        )
-                                    }
-                                    powershell(script: "docker ps --no-trunc --filter \"ancestor=${dockerImageForTesting.imageName()}\" --format \"{{.Names}}\"", returnStatus: true)
                                 }
                             }
                         }
